requirements.txt
requests
rich
psutil
winshell
PyPDF2
GPUtil
pyflakes
pillow
pytesseract
pyperclip
pandas
pyautogui
torch
diffusers
transformers
accelerate
customtkinter
openai
python-dotenv
